Extract asset metadata types
