+++ Sources/Player/NowPlaying.swift
@@ -29,19 +29,6 @@ public struct NowPlayingMetadata {
     }
 }
 
-public protocol AssetMetadata {
-    func nowPlayingMetadata() -> NowPlayingMetadata
-}
-
-public struct EmptyAssetMetadata: AssetMetadata {
-    // Cannot be instantiated. Only meant to be used as a type in signatures and where clauses.
-    private init() {}
-
-    public func nowPlayingMetadata() -> NowPlayingMetadata {
-        .init()
-    }
-}
-
 enum NowPlaying {
     typealias Info = [String: Any]
